requirements.txt
asgiref==3.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-allauth==65.9.0
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-jsonfield-backport==1.0.5
django-modeltranslation==0.19.15
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
jsonfield==3.2.0
oauthlib==3.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.26.3
pyparsing==3.2.3
python3-openid==3.2.0
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
